feat: product editing functionality

+++ components/produtos/DeleteProductDialog.tsx
@@ -15,8 +15,9 @@ export default function DeleteProductDialog(productInput: DeleteProductDialogPro
     const { updateKey } = useContext(ProductModalFormContext);
 
     function deleteProduct() {
+        if(id !== undefined)
         productApi.delete(id)
-        .then(res => successToast(descricao + " deletado com sucesso"!))
+        .then(res => successToast(descricao + " deletado com sucesso!"))
         .catch(error => errorToast(error.message))
         .finally(() => updateKey());
     }
